Ionic Week2 - Einde oefening 3 popups loading spinner gestures

+++ conFusion/www/js/controllers.js
@@ -219,18 +219,28 @@ angular.module('conFusion.controllers', [])
     $scope.baseURL = baseURL;
 
   }])
-  .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate',
-    function($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+  .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout',
+    function($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
     $scope.baseURL = baseURL;
     $scope.shouldShowDelegate = false;
+    $ionicLoading.show({
+      template: '<ion-spinner></ion-spinner> Loading...'
+    });
+
     $scope.favorites = favoriteFactory.getFavorites();
 
       $scope.dishes = menuFactory.getDishes().query(
         function (response) {
           $scope.dishes = response;
+          $timeout(function() {
+            $ionicLoading.hide();
+          }, 1000);
         },
         function (response) {
           $scope.message = "Error: " + response.status + " " + response.statusText;
+          $timeout(function() {
+            $ionicLoading.hide();
+          }, 1000);
         });
       console.log($scope.dishes, $scope.favorites);
 
@@ -241,7 +251,19 @@ angular.module('conFusion.controllers', [])
 
       $scope.deleteFavorite = function (index) {
 
-        favoriteFactory.deleteFromFavorites(index);
+        var confirmPopup = $ionicPopup.confirm({
+          title: 'Confirm Delete',
+          template: 'Zijde zeker?'
+        });
+        confirmPopup.then(function(res) {
+          if (res) {
+            console.log('OK to delete');
+            favoriteFactory.deleteFromFavorites(index);
+          } else  {
+            console.log('Cancelled the delete');
+          }
+        });
+
         $scope.shouldShowDelete = false;
 
       }
